made changes in login page and register page

+++ src/main/java/com/bank/model/CurrentAccount.java
@@ -19,7 +19,7 @@ public class CurrentAccount {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer id;
-	private Integer accountNumber;
+	private Long accountNumber;
 	@ManyToOne
 	private User user;
 }
